In this PR I've tried to speedup the distinct command by optimizing the simulated annealing process and the rearranging of colors.

Non scientific benchmarks show that in fact the proposed changes greatly improve performance. For example, now I'm able to run `pastel distinct 200 | cat` in <1 sec while without these changes it would have taken ~25secs on my machine. However, on small inputs the boost is not as big; in fact `pastel distinct 20 | cat` now runs in ~0.2secs vs ~0.4secs which is still better, but still a bit slow imho.

The simulated annealing was optimized mainly by avoiding to recalculate the distances from scratch every time , but by only recalculating the distances to the color that changed. This is the main contributing factor to the speedup.

After having optimized that, `cargo flamegraph` showed that a _bunch_ of time was spent inside the `rearrange_sequence` function. A first optimization was to simply switch to using `sort_by_cached_key` because calculating the key at each comparison is really too expensive. Then,  I went ahead and removed the sorting altogether because I think there was no need for it and introduced a cache for the minimum distances.

At this point I don't see any other low hanging fruits and I'm likely to stop here as I'm ok with the results.

Let me know what you think.

(also, your implementation of simulated annealing was really cool! simulated annealing is one of my favourite algorithms :smile:  )