@sharkdp 2019-09-01
Thank you very much for working on this!

The speed improvements are impressive. Unfortunately, the behavior isn't quite the same and I can not reproduce the same results. This is a bit difficult to measure, as there is randomness involved, but you can try the following: Run `pastel distinct -v 6` and take a look at the optimized D_min score. For the old version, it should consistently be around 103.2. With your version, it also reaches 103.2 sometimes, but it also yields results which are much worse.

10 runs of the old version:
```
103.23
103.23
103.22
103.22
103.23
103.22
103.23
103.23
103.19
103.22
```

10 runs of the new version:
```
98.90
103.23
103.22
98.90
103.22
103.22
98.90
94.28
98.31
93.48
```

What I can also observe from running `pastel distinct` with the `-v`/`--verbose` option is that the `D_mean` score is excessively high in your version. It looks like a some `scalar::MAX` values accidentally made it into the computation for the mean.

I will add some inline comments and see if I can find any mistakes.

Thanks!