@sharkdp 2019-09-01
> After having optimized that, `cargo flamegraph` showed that a _bunch_ of time was spent inside the `rearrange_sequence` function. A first optimization was to simply switch to using `sort_by_cached_key` because calculating the key at each comparison is really too expensive. Then, I went ahead and removed the sorting altogether because I think there was no need for it and introduced a cache for the minimum distances.

Using `sort_by_cached_key` sounds very reasonable, but I don't understand the second part. What do you mean by "there was no need for it"? I didn't really understand your new implementation when taking a first look, but I can try to take a closer look again.