@sharkdp 2019-09-01
Thank you for the clarification. I somehow missed that the outer loop was still there and got confused as to how that still works :smile:. Yes, that makes sense! Thank you for the clarification.

If I understand it correctly, your changes brings the complexity from *O(N² log N)* down to *O(N²)*.

> Moreover at each step we can be clever and avoid recomputing the minimal distance from each element in the "right" part to each element in the "left" one because we know which element we inserted in the previous run of the loop and so we can update the previous differences with only the new value.

This part I still don't quite understand. We did that before as well, right? I only called `sort_unstable_by_key` on the `right` part. In your implementation, wouldn't it be the same if we just used `min_by_key` on the right part?