@d-dorazio 2019-09-01
> If I understand it correctly, your changes brings the complexity from O(N² log N) down to O(N²).

This is correct if we ignore how the calculation of the keys was done. Previously for each element in "right" it was iterating over each element in "left" to find the maximum which I believe is O(N²) (but I never understood big O :) ) so the final complexity of the previous algorithms (I think) was O(__N²__ N² log N) but now it's O(N²) because this calculation is cached and then updated at each step of the outer loop.