@sharkdp 2019-09-01
Ah, right. That explains why it was so slow for larger N :smile: