@sharkdp 2019-09-19
Released in [v0.6.0](https://github.com/sharkdp/pastel/releases).